requirements.txt
# File: requirements.txt
requests>=2.31.0
pydantic>=2.5.0
pandas>=2.0.0
python-dateutil>=2.8.2
tzdata>=2023.3
